[feat] : Home, Profile 네비게이션 연결
