+++ app/build.gradle.kts
@@ -33,6 +33,7 @@ dependencies {
 
     // feature
     implementation(projects.feature.profile)
+    implementation(projects.feature.home)
 //    implementation(projects.feature.match)
 
 //    implementation(libs.coil.core)
